🐛  Big refactor on API results
I had two big mistakes in the first version of the client:

* I was mistakenly thinking they structured their data in a specific
  way, and they didn't, which lead to me missing data. Now, it is there
  in it's complete glory.
* I was snakecasing the Query params, and that was modifying them in
  ways I did not want. Just drop that.

At this point, the cods are right, and the returned data isn't as clean,
but it's what Persona returns, and that's got to be better in the
long-run. Al the tests run and pass, which is where the doc responses
come from, so that's good, and the lookup of the Verification from the
Inquiry now is clearly possible.