🔩  Had to pull Persona config to client
We really needed to be able to allow the caller to place all the
template ids in the client on creation so that we would not have to pass
any of them into the calls to get what we wanted. Of course, you still
can, and the parameters will override the client config, but it will be
nice to be able to default all the functions with the right templates.